requirements.txt
Django==4.2.6
PyYAML
djangorestframework
requests

# Development
debugpy
